requirements.txt
scipy
Flask==1.1.2
joblib==1.0.0
lxml==4.5.1
numpy==1.19.5
PySastrawi==1.2.0
requests==2.23.0
requests-html==0.10.0
scikit-learn==0.24.0
tqdm==4.46.0
urllib3==1.25.9
virtualenv==20.0.27
w3lib==1.22.0
websockets==8.1
Werkzeug==1.0.1